# feat: Performance-Optimierung - Entity Management und Kollisionserkennung