… PERFORMANCE_OPTIMIZATIONS.md
… index.html
… js/performance/spatial-manager.js
… test-performance.html